README update and class file shifted
